se pueden editar los estados de pedido en el desktop

+++ TPIDesktop/src/main/java/Persistencia/CarritoJpaController.java
@@ -4,17 +4,20 @@
  */
 package Persistencia;
 
-import Modelo.Carrito;
-import Persistencia.exceptions.NonexistentEntityException;
 import java.io.Serializable;
+import javax.persistence.Query;
+import javax.persistence.EntityNotFoundException;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
+import Modelo.Articulo;
+import Modelo.Carrito;
+import Persistencia.exceptions.IllegalOrphanException;
+import Persistencia.exceptions.NonexistentEntityException;
+import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
 import javax.persistence.Persistence;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
 
 /**
  *
@@ -25,7 +28,6 @@ public class CarritoJpaController implements Serializable {
     public CarritoJpaController(EntityManagerFactory emf) {
         this.emf = emf;
     }
-
     public CarritoJpaController() {
         emf = Persistence.createEntityManagerFactory("TPIPU");
     }
@@ -36,11 +38,29 @@ public class CarritoJpaController implements Serializable {
     }
 
     public void create(Carrito carrito) {
+        if (carrito.getArticulos() == null) {
+            carrito.setArticulos(new ArrayList<Articulo>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            List<Articulo> attachedArticulos = new ArrayList<Articulo>();
+            for (Articulo articulosArticuloToAttach : carrito.getArticulos()) {
+                articulosArticuloToAttach = em.getReference(articulosArticuloToAttach.getClass(), articulosArticuloToAttach.getId_articulo());
+                attachedArticulos.add(articulosArticuloToAttach);
+            }
+            carrito.setArticulos(attachedArticulos);
             em.persist(carrito);
+            for (Articulo articulosArticulo : carrito.getArticulos()) {
+                Carrito oldCarritoOfArticulosArticulo = articulosArticulo.getCarrito();
+                articulosArticulo.setCarrito(carrito);
+                articulosArticulo = em.merge(articulosArticulo);
+                if (oldCarritoOfArticulosArticulo != null) {
+                    oldCarritoOfArticulosArticulo.getArticulos().remove(articulosArticulo);
+                    oldCarritoOfArticulosArticulo = em.merge(oldCarritoOfArticulosArticulo);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -49,12 +69,45 @@ public class CarritoJpaController implements Serializable {
         }
     }
 
-    public void edit(Carrito carrito) throws NonexistentEntityException, Exception {
+    public void edit(Carrito carrito) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Carrito persistentCarrito = em.find(Carrito.class, carrito.getId());
+            List<Articulo> articulosOld = persistentCarrito.getArticulos();
+            List<Articulo> articulosNew = carrito.getArticulos();
+            List<String> illegalOrphanMessages = null;
+            for (Articulo articulosOldArticulo : articulosOld) {
+                if (!articulosNew.contains(articulosOldArticulo)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Articulo " + articulosOldArticulo + " since its carrito field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<Articulo> attachedArticulosNew = new ArrayList<Articulo>();
+            for (Articulo articulosNewArticuloToAttach : articulosNew) {
+                articulosNewArticuloToAttach = em.getReference(articulosNewArticuloToAttach.getClass(), articulosNewArticuloToAttach.getId_articulo());
+                attachedArticulosNew.add(articulosNewArticuloToAttach);
+            }
+            articulosNew = attachedArticulosNew;
+            carrito.setArticulos(articulosNew);
             carrito = em.merge(carrito);
+            for (Articulo articulosNewArticulo : articulosNew) {
+                if (!articulosOld.contains(articulosNewArticulo)) {
+                    Carrito oldCarritoOfArticulosNewArticulo = articulosNewArticulo.getCarrito();
+                    articulosNewArticulo.setCarrito(carrito);
+                    articulosNewArticulo = em.merge(articulosNewArticulo);
+                    if (oldCarritoOfArticulosNewArticulo != null && !oldCarritoOfArticulosNewArticulo.equals(carrito)) {
+                        oldCarritoOfArticulosNewArticulo.getArticulos().remove(articulosNewArticulo);
+                        oldCarritoOfArticulosNewArticulo = em.merge(oldCarritoOfArticulosNewArticulo);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -72,7 +125,7 @@ public class CarritoJpaController implements Serializable {
         }
     }
 
-    public void destroy(int id) throws NonexistentEntityException {
+    public void destroy(int id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -83,6 +136,17 @@ public class CarritoJpaController implements Serializable {
                 carrito.getId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The carrito with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<Articulo> articulosOrphanCheck = carrito.getArticulos();
+            for (Articulo articulosOrphanCheckArticulo : articulosOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Carrito (" + carrito + ") cannot be destroyed since the Articulo " + articulosOrphanCheckArticulo + " in its articulos field has a non-nullable carrito field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(carrito);
             em.getTransaction().commit();
@@ -138,5 +202,5 @@ public class CarritoJpaController implements Serializable {
             em.close();
         }
     }
-
+    
 }
