+++ TPIDesktop/src/main/java/Persistencia/PersonaJpaController.java
@@ -4,17 +4,20 @@
  */
 package Persistencia;
 
-import Modelo.Persona;
-import Persistencia.exceptions.NonexistentEntityException;
 import java.io.Serializable;
+import javax.persistence.Query;
+import javax.persistence.EntityNotFoundException;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
+import Modelo.Pedido;
+import Modelo.Persona;
+import Persistencia.exceptions.IllegalOrphanException;
+import Persistencia.exceptions.NonexistentEntityException;
+import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
-import javax.persistence.EntityNotFoundException;
 import javax.persistence.Persistence;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
 
 /**
  *
@@ -25,7 +28,6 @@ public class PersonaJpaController implements Serializable {
     public PersonaJpaController(EntityManagerFactory emf) {
         this.emf = emf;
     }
-
     public PersonaJpaController() {
         emf = Persistence.createEntityManagerFactory("TPIPU");
     }
@@ -36,11 +38,29 @@ public class PersonaJpaController implements Serializable {
     }
 
     public void create(Persona persona) {
+        if (persona.getPedidos() == null) {
+            persona.setPedidos(new ArrayList<Pedido>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            List<Pedido> attachedPedidos = new ArrayList<Pedido>();
+            for (Pedido pedidosPedidoToAttach : persona.getPedidos()) {
+                pedidosPedidoToAttach = em.getReference(pedidosPedidoToAttach.getClass(), pedidosPedidoToAttach.getId());
+                attachedPedidos.add(pedidosPedidoToAttach);
+            }
+            persona.setPedidos(attachedPedidos);
             em.persist(persona);
+            for (Pedido pedidosPedido : persona.getPedidos()) {
+                Persona oldUsuarioOfPedidosPedido = pedidosPedido.getUsuario();
+                pedidosPedido.setUsuario(persona);
+                pedidosPedido = em.merge(pedidosPedido);
+                if (oldUsuarioOfPedidosPedido != null) {
+                    oldUsuarioOfPedidosPedido.getPedidos().remove(pedidosPedido);
+                    oldUsuarioOfPedidosPedido = em.merge(oldUsuarioOfPedidosPedido);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -49,12 +69,45 @@ public class PersonaJpaController implements Serializable {
         }
     }
 
-    public void edit(Persona persona) throws NonexistentEntityException, Exception {
+    public void edit(Persona persona) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Persona persistentPersona = em.find(Persona.class, persona.getId());
+            List<Pedido> pedidosOld = persistentPersona.getPedidos();
+            List<Pedido> pedidosNew = persona.getPedidos();
+            List<String> illegalOrphanMessages = null;
+            for (Pedido pedidosOldPedido : pedidosOld) {
+                if (!pedidosNew.contains(pedidosOldPedido)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Pedido " + pedidosOldPedido + " since its usuario field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<Pedido> attachedPedidosNew = new ArrayList<Pedido>();
+            for (Pedido pedidosNewPedidoToAttach : pedidosNew) {
+                pedidosNewPedidoToAttach = em.getReference(pedidosNewPedidoToAttach.getClass(), pedidosNewPedidoToAttach.getId());
+                attachedPedidosNew.add(pedidosNewPedidoToAttach);
+            }
+            pedidosNew = attachedPedidosNew;
+            persona.setPedidos(pedidosNew);
             persona = em.merge(persona);
+            for (Pedido pedidosNewPedido : pedidosNew) {
+                if (!pedidosOld.contains(pedidosNewPedido)) {
+                    Persona oldUsuarioOfPedidosNewPedido = pedidosNewPedido.getUsuario();
+                    pedidosNewPedido.setUsuario(persona);
+                    pedidosNewPedido = em.merge(pedidosNewPedido);
+                    if (oldUsuarioOfPedidosNewPedido != null && !oldUsuarioOfPedidosNewPedido.equals(persona)) {
+                        oldUsuarioOfPedidosNewPedido.getPedidos().remove(pedidosNewPedido);
+                        oldUsuarioOfPedidosNewPedido = em.merge(oldUsuarioOfPedidosNewPedido);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -72,7 +125,7 @@ public class PersonaJpaController implements Serializable {
         }
     }
 
-    public void destroy(int id) throws NonexistentEntityException {
+    public void destroy(int id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -83,6 +136,17 @@ public class PersonaJpaController implements Serializable {
                 persona.getId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The persona with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<Pedido> pedidosOrphanCheck = persona.getPedidos();
+            for (Pedido pedidosOrphanCheckPedido : pedidosOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Persona (" + persona + ") cannot be destroyed since the Pedido " + pedidosOrphanCheckPedido + " in its pedidos field has a non-nullable usuario field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(persona);
             em.getTransaction().commit();
@@ -138,5 +202,5 @@ public class PersonaJpaController implements Serializable {
             em.close();
         }
     }
-
+    
 }
