+++ TPIDesktop/src/main/java/Persistencia/ProductoJpaController.java
@@ -4,17 +4,21 @@
  */
 package Persistencia;
 
-import Modelo.Producto;
-import Persistencia.exceptions.NonexistentEntityException;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
-import javax.persistence.Persistence;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import Modelo.Articulo;
+import java.util.ArrayList;
+import java.util.List;
+import Modelo.Detalle_de_pedido;
+import Modelo.Producto;
+import Persistencia.exceptions.IllegalOrphanException;
+import Persistencia.exceptions.NonexistentEntityException;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
 
 /**
  *
@@ -25,7 +29,8 @@ public class ProductoJpaController implements Serializable {
     public ProductoJpaController(EntityManagerFactory emf) {
         this.emf = emf;
     }
-    public ProductoJpaController(){
+    
+    public ProductoJpaController() {
         emf = Persistence.createEntityManagerFactory("TPIPU");
     }
     private EntityManagerFactory emf = null;
@@ -35,11 +40,47 @@ public class ProductoJpaController implements Serializable {
     }
 
     public void create(Producto producto) {
+        if (producto.getArticulos() == null) {
+            producto.setArticulos(new ArrayList<Articulo>());
+        }
+        if (producto.getDetalle_de_pedido() == null) {
+            producto.setDetalle_de_pedido(new ArrayList<Detalle_de_pedido>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            List<Articulo> attachedArticulos = new ArrayList<Articulo>();
+            for (Articulo articulosArticuloToAttach : producto.getArticulos()) {
+                articulosArticuloToAttach = em.getReference(articulosArticuloToAttach.getClass(), articulosArticuloToAttach.getId_articulo());
+                attachedArticulos.add(articulosArticuloToAttach);
+            }
+            producto.setArticulos(attachedArticulos);
+            List<Detalle_de_pedido> attachedDetalle_de_pedido = new ArrayList<Detalle_de_pedido>();
+            for (Detalle_de_pedido detalle_de_pedidoDetalle_de_pedidoToAttach : producto.getDetalle_de_pedido()) {
+                detalle_de_pedidoDetalle_de_pedidoToAttach = em.getReference(detalle_de_pedidoDetalle_de_pedidoToAttach.getClass(), detalle_de_pedidoDetalle_de_pedidoToAttach.getId());
+                attachedDetalle_de_pedido.add(detalle_de_pedidoDetalle_de_pedidoToAttach);
+            }
+            producto.setDetalle_de_pedido(attachedDetalle_de_pedido);
             em.persist(producto);
+            for (Articulo articulosArticulo : producto.getArticulos()) {
+                Producto oldProductoOfArticulosArticulo = articulosArticulo.getProducto();
+                articulosArticulo.setProducto(producto);
+                articulosArticulo = em.merge(articulosArticulo);
+                if (oldProductoOfArticulosArticulo != null) {
+                    oldProductoOfArticulosArticulo.getArticulos().remove(articulosArticulo);
+                    oldProductoOfArticulosArticulo = em.merge(oldProductoOfArticulosArticulo);
+                }
+            }
+            for (Detalle_de_pedido detalle_de_pedidoDetalle_de_pedido : producto.getDetalle_de_pedido()) {
+                Producto oldProductoOfDetalle_de_pedidoDetalle_de_pedido = detalle_de_pedidoDetalle_de_pedido.getProducto();
+                detalle_de_pedidoDetalle_de_pedido.setProducto(producto);
+                detalle_de_pedidoDetalle_de_pedido = em.merge(detalle_de_pedidoDetalle_de_pedido);
+                if (oldProductoOfDetalle_de_pedidoDetalle_de_pedido != null) {
+                    oldProductoOfDetalle_de_pedidoDetalle_de_pedido.getDetalle_de_pedido().remove(detalle_de_pedidoDetalle_de_pedido);
+                    oldProductoOfDetalle_de_pedidoDetalle_de_pedido = em.merge(oldProductoOfDetalle_de_pedidoDetalle_de_pedido);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -48,12 +89,71 @@ public class ProductoJpaController implements Serializable {
         }
     }
 
-    public void edit(Producto producto) throws NonexistentEntityException, Exception {
+    public void edit(Producto producto) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Producto persistentProducto = em.find(Producto.class, producto.getId());
+            List<Articulo> articulosOld = persistentProducto.getArticulos();
+            List<Articulo> articulosNew = producto.getArticulos();
+            List<Detalle_de_pedido> detalle_de_pedidoOld = persistentProducto.getDetalle_de_pedido();
+            List<Detalle_de_pedido> detalle_de_pedidoNew = producto.getDetalle_de_pedido();
+            List<String> illegalOrphanMessages = null;
+            for (Articulo articulosOldArticulo : articulosOld) {
+                if (!articulosNew.contains(articulosOldArticulo)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Articulo " + articulosOldArticulo + " since its producto field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<Articulo> attachedArticulosNew = new ArrayList<Articulo>();
+            for (Articulo articulosNewArticuloToAttach : articulosNew) {
+                articulosNewArticuloToAttach = em.getReference(articulosNewArticuloToAttach.getClass(), articulosNewArticuloToAttach.getId_articulo());
+                attachedArticulosNew.add(articulosNewArticuloToAttach);
+            }
+            articulosNew = attachedArticulosNew;
+            producto.setArticulos(articulosNew);
+            List<Detalle_de_pedido> attachedDetalle_de_pedidoNew = new ArrayList<Detalle_de_pedido>();
+            for (Detalle_de_pedido detalle_de_pedidoNewDetalle_de_pedidoToAttach : detalle_de_pedidoNew) {
+                detalle_de_pedidoNewDetalle_de_pedidoToAttach = em.getReference(detalle_de_pedidoNewDetalle_de_pedidoToAttach.getClass(), detalle_de_pedidoNewDetalle_de_pedidoToAttach.getId());
+                attachedDetalle_de_pedidoNew.add(detalle_de_pedidoNewDetalle_de_pedidoToAttach);
+            }
+            detalle_de_pedidoNew = attachedDetalle_de_pedidoNew;
+            producto.setDetalle_de_pedido(detalle_de_pedidoNew);
             producto = em.merge(producto);
+            for (Articulo articulosNewArticulo : articulosNew) {
+                if (!articulosOld.contains(articulosNewArticulo)) {
+                    Producto oldProductoOfArticulosNewArticulo = articulosNewArticulo.getProducto();
+                    articulosNewArticulo.setProducto(producto);
+                    articulosNewArticulo = em.merge(articulosNewArticulo);
+                    if (oldProductoOfArticulosNewArticulo != null && !oldProductoOfArticulosNewArticulo.equals(producto)) {
+                        oldProductoOfArticulosNewArticulo.getArticulos().remove(articulosNewArticulo);
+                        oldProductoOfArticulosNewArticulo = em.merge(oldProductoOfArticulosNewArticulo);
+                    }
+                }
+            }
+            for (Detalle_de_pedido detalle_de_pedidoOldDetalle_de_pedido : detalle_de_pedidoOld) {
+                if (!detalle_de_pedidoNew.contains(detalle_de_pedidoOldDetalle_de_pedido)) {
+                    detalle_de_pedidoOldDetalle_de_pedido.setProducto(null);
+                    detalle_de_pedidoOldDetalle_de_pedido = em.merge(detalle_de_pedidoOldDetalle_de_pedido);
+                }
+            }
+            for (Detalle_de_pedido detalle_de_pedidoNewDetalle_de_pedido : detalle_de_pedidoNew) {
+                if (!detalle_de_pedidoOld.contains(detalle_de_pedidoNewDetalle_de_pedido)) {
+                    Producto oldProductoOfDetalle_de_pedidoNewDetalle_de_pedido = detalle_de_pedidoNewDetalle_de_pedido.getProducto();
+                    detalle_de_pedidoNewDetalle_de_pedido.setProducto(producto);
+                    detalle_de_pedidoNewDetalle_de_pedido = em.merge(detalle_de_pedidoNewDetalle_de_pedido);
+                    if (oldProductoOfDetalle_de_pedidoNewDetalle_de_pedido != null && !oldProductoOfDetalle_de_pedidoNewDetalle_de_pedido.equals(producto)) {
+                        oldProductoOfDetalle_de_pedidoNewDetalle_de_pedido.getDetalle_de_pedido().remove(detalle_de_pedidoNewDetalle_de_pedido);
+                        oldProductoOfDetalle_de_pedidoNewDetalle_de_pedido = em.merge(oldProductoOfDetalle_de_pedidoNewDetalle_de_pedido);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -71,7 +171,7 @@ public class ProductoJpaController implements Serializable {
         }
     }
 
-    public void destroy(int id) throws NonexistentEntityException {
+    public void destroy(int id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -82,6 +182,22 @@ public class ProductoJpaController implements Serializable {
                 producto.getId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The producto with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<Articulo> articulosOrphanCheck = producto.getArticulos();
+            for (Articulo articulosOrphanCheckArticulo : articulosOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Producto (" + producto + ") cannot be destroyed since the Articulo " + articulosOrphanCheckArticulo + " in its articulos field has a non-nullable producto field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<Detalle_de_pedido> detalle_de_pedido = producto.getDetalle_de_pedido();
+            for (Detalle_de_pedido detalle_de_pedidoDetalle_de_pedido : detalle_de_pedido) {
+                detalle_de_pedidoDetalle_de_pedido.setProducto(null);
+                detalle_de_pedidoDetalle_de_pedido = em.merge(detalle_de_pedidoDetalle_de_pedido);
             }
             em.remove(producto);
             em.getTransaction().commit();
